+++ src/WorldView.cpp
@@ -29,8 +29,8 @@ bool WorldView::init()
 		ts.init(std::get<1>(worldTileSet), 
 			    std::get<2>(worldTileSet));
 
-		uint16_t worldMapWidth = world.worldMap.size();
-		uint16_t worldMapHeight = world.worldMap[0].size();
+		uint32_t worldMapWidth = world.worldMap.size();
+		uint32_t worldMapHeight = world.worldMap[0].size();
 
 		biomeDisplay.init(0,
 			      		  0,
@@ -76,10 +76,9 @@ void WorldView::redrawWorld()
 	if (world.getFirstTime())
 	{
 		std::vector<Tile>* tile_array = biomeDisplay.getTiles();
-		uint16_t worldMapWidth = world.worldMap.size();
-		//uint16_t worldMapHeight = world.worldMap[0].size();
-
-		for (uint16_t i = 0; i< (*tile_array).size(); i++)
+		uint32_t worldMapWidth = world.worldMap.size();
+		
+		for (uint32_t i = 0; i< (*tile_array).size(); i++)
 		{
 			(*tile_array)[i] = world.getTile(i%worldMapWidth, i/worldMapWidth);
 		}
@@ -103,14 +102,14 @@ DefinedViews WorldView::handleKeyPress(ALLEGRO_EVENT* ev)
 	DefinedViews dv = DefinedViews::WORLD_VIEW;
 	std::vector<Tile>* tile_array = biomeDisplay.getTiles();
 	World& world = model->getWorld();
-	uint16_t worldMapWidth = world.worldMap.size();
-	uint16_t worldMapHeight = world.worldMap[0].size();
+	uint32_t worldMapWidth = world.worldMap.size();
+	uint32_t worldMapHeight = world.worldMap[0].size();
 
 	Player& player = model->getPlayer();
-	uint16_t playerWorldX = player.getWorldX();
-	uint16_t playerWorldY = player.getWorldY();
-	uint16_t newX = playerWorldX;
-	uint16_t newY = playerWorldY;
+	uint32_t playerWorldX = player.getWorldX();
+	uint32_t playerWorldY = player.getWorldY();
+	uint32_t newX = playerWorldX;
+	uint32_t newY = playerWorldY;
 	bool move = false;
 	bool switchView = false;
 	
@@ -213,8 +212,8 @@ void WorldView::updateBiomePane()
 
 	World& world = model->getWorld();
 	Player& player = model->getPlayer();
-	uint16_t playerWorldX = player.getWorldX();
-	uint16_t playerWorldY = player.getWorldY();
+	uint32_t playerWorldX = player.getWorldX();
+	uint32_t playerWorldY = player.getWorldY();
 
 	biomeInfo.write("                              Biome Info");
 	biomeInfo.write(" ");
@@ -254,9 +253,9 @@ std::vector<Displayable*>& WorldView::draw()
 	{
 		std::vector<Tile>* tile_array = biomeDisplay.getTiles();
 		Player& player = model->getPlayer();
-		uint16_t playerX = player.getWorldX();
-		uint16_t playerY = player.getWorldY();
-		uint16_t worldMapWidth = model->getWorld().getXSize();
+		uint32_t playerX = player.getWorldX();
+		uint32_t playerY = player.getWorldY();
+		uint32_t worldMapWidth = model->getWorld().getXSize();
 		Tile& t = (*tile_array)[playerX+playerY*worldMapWidth];
 		/* draw the player */
 		t.setIndex(player.getChar());
